+++ tempest_fastapi_sdk/__init__.py
@@ -154,7 +154,7 @@ from tempest_fastapi_sdk.webpush import (
     WebPushSubscriptionSchema,
 )
 
-__version__: str = "0.19.1"
+__version__: str = "0.19.2"
 
 __all__: list[str] = [
     "CEP",